pyproject.toml
[project]
name = "vaje"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
    "dvc[s3]>=3.67.1",
    "evidently>=0.7.1,<0.8.0",
    "great_expectations==0.18.21",
    "lxml>=5.3.1,<6.0.0",
    "numpy>=1.22.4,<2.0.0",
    "pandas>=2.2.3,<3.0.0",
    "pyyaml>=6.0,<7.0",
    "requests>=2.32.3,<3.0.0",
    "scikit-learn>=1.6.1,<2.0.0",
    "tensorflow>=2.19.0,<3.0.0",
]